--- src/components/MovieCard/MovieCard.js
@@ -1,6 +1,0 @@
-import React from 'react';
-import './MovieCard.scss';
-
-export const MovieCard = () => (
-  <>Put the card here</>
-);
