--- src/components/MoviesList/MoviesList.js
@@ -1,6 +1,0 @@
-import React from 'react';
-import './MoviesList.scss';
-
-export const MoviesList = () => (
-  <>Put the list here</>
-);
