+++ src/components/MoviesList/MoviesList.jsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import './MoviesList.scss';
+import PropTypes from 'prop-types';
+import { MovieCard } from '../MovieCard';
+
+export const MoviesList = ({ movies }) => (
+  <>
+    <div className="movies" key="movies.imdbId">
+      {movies.map(movie => (
+        <MovieCard {...movie} />
+      ))}
+    </div>
+  </>
+);
+
+MoviesList.propTypes = {
+  movies: PropTypes.arrayOf.isRequired,
+};
