+++ src/components/MoviesList/MoviesList.scss
@@ -1,1 +1,5 @@
-
+.movies {
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(300px, 1fr));
+  gap: 20px;
+}
